requirements.txt
chatbotAI==0.3.1.3
cx_Freeze==7.2.7
cx-Logging==3.2.1
ffmpeg==1.4
ffmpeg_python==0.2.0
fuzzywuzzy==0.18.0
geocoder==1.38.1
lief==0.15.1
loguru==0.6.0
notify-py==0.3.43
num2words==0.5.10
numpy==2.0.2
pip==24.3.1
pvporcupine==1.8.7
PyAudio==0.2.14
pykeepass==4.1.0.post1
pynput==1.7.7
pyowm==3.3.0
python_dateutil==2.9.0.post0
pyttsx3==2.98
pytz==2024.2
PyYAML==6.0.2
Requests==2.32.3
Resemblyzer==0.1.4
scipy==1.14.1
sentencepiece==0.2.0
sounddevice==0.5.1
soundfile==0.12.1
text2numde==1.0
tinydb==4.8.2
transformers==4.46.1
vosk==0.3.45
wikipedia==1.4.0
words2num==0.2.3
wxPython==4.2.2